requirements.txt
affine==2.4.0
ansi2html==1.9.1
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.4
boto3==1.34.112
botocore==1.34.112
Brotli==1.1.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.0.1
comm==0.2.1
contourpy==1.2.0
cramjam==2.8.1
cycler==0.12.1
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-leaflet==1.0.15
dash-resizable-panels==0.1.0
dash-svg==0.0.12
dash-table==5.0.0
dash_daq==0.5.0
dash_deck==0.0.1
dask==2024.2.0
datashader==0.16.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
fastparquet==2024.2.0
fiona==1.9.5
Flask==3.0.1
Flask-Caching==2.3.0
Flask-Compress==1.15
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.2.0
geopandas==0.14.2
holoviews==1.18.3
hvplot==0.9.2
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
linkify-it-py==2.0.3
llvmlite==0.42.0
locket==1.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
multipledispatch==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.0.8
notebook_shim==0.2.3
numba==0.59.0
numpy==1.26.3
osmnx==1.9.1
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
panel==1.3.8
param==2.0.2
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.5.0
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.4
pyviz_comms==3.0.1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
referencing==0.33.0
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rioxarray==0.15.5
rpds-py==0.17.1
s3transfer==0.10.1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
xarray==2024.1.1
xyzservices==2023.10.1
yarg==0.1.9
zipp==3.17.0
zstandard==0.22.0
